# Conflicts:
#	WebContent/LoginPage.html

+++ WebContent/scripts/function.js
@@ -70,15 +70,14 @@ function edit_profile(){ // checked
 {
     //$("registerbutton").attr({"disabled":"disabled"});
     var flag = false;
-    username = $("#Usernane").val();
-    pwd =  $("#password").val();
+    nickname = $("#nickname").val();
+   // pwd =  $("#password").val();
 	wechatid = $("#wechatId").val();
 	email = $("#Email").val();
 	/*phonenumber = $("PhoneNumber").val();*/
 	oldpwd = $("#useroldpassword").val();
 	newpwd = $("#userpassword").val();
 	confirmpwd = $("#confirmPassword").val();
-	
     var reg = new RegExp("^\\w*$"); // match/cleans the entry (spaces,...), for verifying entry
 	var filter = /^(([^<>()[]\.,;:s@"]+(.[^<>()[]\.,;:s@"]+)*)|(".+"))@(([[0-9]{1,3}.[0-9]{1,3}.[0-9]{1,3}.[0-9]{1,3}])|(([a-zA-Z-0-9]+.)+[a-zA-Z]{2,}))$/; // filter for email adress
 
@@ -108,8 +107,8 @@ function edit_profile(){ // checked
         $("#userpassword").focus();
         return false;
 	}
-    
-    data = {'wechat_id':wechatid,'new_pw_hash':$.md5(newpwd), 'email':email/*, 'phone_number':phonenumber*/}; //creating json file
+   	
+    data = {'u_id': $.cookie('u_id'),'u_token': $.cookie('u_token'),'u_name': $.cookie('u_name'), 'u_full_name': nickname ,'u_wechat_id':wechatid,'new_pw_hash':$.md5(newpwd), 'u_email_address':email/*, 'phone_number':phonenumber*/}; //creating json file
     //alert(data.wechat_id);
 	
 	
@@ -138,16 +137,16 @@ function edit_profile(){ // checked
         type : "post",
         data : data,
         async: false,
-        url : "/ccpx/user/edit_profile", ///////////////////////// won't work because of this file
+        url : "user/modify", ///////////////////////// won't work because of this file
         success : function(result){
             if (result.errno==0) { // parameter in their response
-                if($.md5(oldpwd) != getCookie("u_pw_hash")){ ///////////////////////////////////////////////get the value of the current password/////////////////////////////////////////////////
+                if($.md5(oldpwd) != $.cookie('u_pw_hash')){ ///////////////////////////////////////////////get the value of the current password/////////////////////////////////////////////////
 					toastr.warning("Your old password is not valid", "Warning");
 					$("#useroldpassword").focus();
 					return false;
-				}else{
+				}else{ 
 					toastr.success(result.rsm.token, "info");
-					location.href="MyAccount.html"
+					location.href="MyAccount.html";
 				}
             }else{
                 toastr.warning(result.err, "Warning:CODE "+result.errno); //pop up
@@ -160,10 +159,11 @@ function edit_profile(){ // checked
     return false;
     
     //toastr.warning(flag.toString(), "DEBUG");
-    $("#loginButton").removeAttr("disabled");
+   // $("#loginButton").removeAttr("disabled");
     return flag;
 }
 }
+
 
 
 
@@ -299,24 +299,24 @@ function get_profile(){
 	**** end test ***/
 	
 	
-	var id = getCookie("u_id");
-	var token = getCookie("u_token");
+	var id = $.cookie('u_id');
+	var token = $.cookie('u_token');
 	$.ajax({
 		type : "get",
 		dataType:"JSON",
-		data : {u_id: id, u_token: token},
+		data : {'u_id': id, 'u_token': token},
 		async: false,
-		url	: "/ccpx/user/myprofile", 
+		url	: "user/myprofile", 
 		success : function(result){
 			if (result.errno==0) { // parameter in their response
                 toastr.success(result.rsm.token, "info");
-				$.cookie('u_pw_hash',result.rsm.u_pw_hash);
 				/*$("#userName").append(result.rsm.u_name);*/
-				$("#wechatId").text(result.rsm.wechatid);
-				$("#Email").text(result.rsm.u_email);
+                $("#nickname").val(result.rsm.u_fullname);
+				$("#wechatId").val(result.rsm.u_wechat_id);
+				$("#Email").val(result.rsm.u_email);
 				/*$("#PhoneNumber").text(result.rsm."");*/
-				$("#userpassword").text(result.rsm.u_pw_hash);
-                location.href="MyAccount.html"
+				//$("#useroldpassword").val(pwd);
+				$.cookie('u_name',result.rsm.u_name);
             }else{
                 toastr.warning(result.err, "Warning:CODE "+result.errno); //pop up
             }
@@ -325,7 +325,7 @@ function get_profile(){
 		error:function(){
 			toastr.error("error", "error");
 		}
-	})
+	});
 }
 
 function getAllSellerInfo() {
@@ -490,15 +490,18 @@ function make_an_offer(){
 }
 
  function logout() {
-	 
-	 var delete_cookie = function(name) {
-		document.cookie = name + '=;expires=Thu, 01 Jan 1970 00:00:01 GMT;';
-	};
-	delete_cookie('u_id');
-	delete_cookie('u_token');
-	location.href="MainPage.html";
+	 if (confirm("Are you sure to Logout ? ")) {  
+	        var delete_cookie = function(name) {
+			document.cookie = name + '=;expires=Thu, 01 Jan 1970 00:00:01 GMT;';
+		    };
+			delete_cookie('u_id');
+			delete_cookie('u_token');
+			delete_cookie('u_name');
+			delete_cookie('u_pw_hash');
+			location.href="MainPage.html";
+	    }  
+		 	
 }
-
